Add react-native-otp-entry package and update OtpScreen with OtpInput component
This commit adds the "react-native-otp-entry" package to the project and updates the "OtpScreen" component to
