--- redM-shift-schedule-app/screens/LogInScreens/LogIn.js
@@ -1,1 +1,0 @@
-import React from 'react';
